+++ driver/web/rest/apis.go
@@ -19,13 +19,18 @@ package rest
 
 import (
 	"content/core"
+	"content/core/model"
 	"encoding/json"
 	"github.com/gorilla/mux"
 	"go.mongodb.org/mongo-driver/bson"
+	"io/ioutil"
 	"log"
 	"net/http"
+	"strconv"
 	"strings"
 )
+
+const maxUploadSize = 15 * 1024 * 1024 // 15 mb
 
 //ApisHandler handles the rest APIs implementation
 type ApisHandler struct {
@@ -113,6 +118,102 @@ func (h ApisHandler) GetStudentGuide(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=utf-8")
 	w.WriteHeader(http.StatusOK)
 	w.Write(data)
+}
+
+// UploadImage Uploads an image to AWS S3
+// @Description Uploads an image to AWS S3
+// @Tags Client
+// @ID AdminUpdateStudentGuide
+// @Param path body string true "path - path within the S3 bucket"
+// @Param width body string false "width - width of the image to resize. If width and height are missing - then the new image will use the original size"
+// @Param height body string false "height - height of the image to resize. If width and height are missing - then the new image will use the original size"
+// @Param quality body string false "quality - quality of the image. Default: 90"
+// @Param fileName body string false "fileName - the uploaded file name"
+// @Accept multipart/form-data
+// @Produce json
+// @Success 200
+// @Security RokwireAuth
+// @Router /image [post]
+func (h ApisHandler) UploadImage(w http.ResponseWriter, r *http.Request) {
+	//validate the image type
+	path := r.PostFormValue("path")
+	if len(path) <= 0 {
+		log.Print("Missing image path\n")
+		http.Error(w, "missing 'path' form param", http.StatusBadRequest)
+		return
+	}
+
+	heightParam := intPostValueFromString(r.PostFormValue("height"))
+	widthParam := intPostValueFromString(r.PostFormValue("width"))
+	qualityParam := intPostValueFromString(r.PostFormValue("quality"))
+	imgSpec := model.ImageSpec{Height: heightParam, Width: widthParam, Quality: qualityParam}
+
+	// validate file size
+	r.Body = http.MaxBytesReader(w, r.Body, maxUploadSize)
+	if err := r.ParseMultipartForm(maxUploadSize); err != nil {
+		log.Print("File is too big\n")
+		http.Error(w, "File is too big", http.StatusBadRequest)
+		return
+	}
+
+	// parse and validate file and post parameters
+	file, fileHeader, err := r.FormFile("fileName")
+	if err != nil {
+		log.Print("Invalid file\n")
+		http.Error(w, "Invalid file", http.StatusBadRequest)
+		return
+	}
+	defer file.Close()
+	fileBytes, err := ioutil.ReadAll(file)
+	if err != nil {
+		log.Print("Invalid file\n")
+		http.Error(w, "Invalid file", http.StatusBadRequest)
+		return
+	}
+
+	// check file type, detectcontenttype only needs the first 512 bytes
+	filetype := http.DetectContentType(fileBytes)
+	switch filetype {
+	case "image/jpeg", "image/jpg":
+	case "image/gif", "image/png":
+	case "image/webp":
+		break
+	default:
+		log.Print("Invalid file type\n")
+		http.Error(w, "Invalid file type", http.StatusBadRequest)
+		return
+	}
+
+	// pass the file to be processed by the use case handler
+	fileName := fileHeader.Filename
+	objectLocation, err := h.app.Services.UploadImage(fileName, filetype, fileBytes, path, imgSpec)
+	if err != nil {
+		log.Printf("Error converting image: %s\n", err)
+		http.Error(w, "Error converting image", http.StatusInternalServerError)
+		return
+	}
+
+	jsonData, err := json.Marshal(objectLocation)
+	if err != nil {
+		log.Println("Error on marshal s3 location data")
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+	w.WriteHeader(http.StatusOK)
+	w.Write(jsonData)
+}
+
+func intPostValueFromString(stringValue string) int {
+	var value int
+	if len(stringValue) > 0 {
+		val, err := strconv.Atoi(stringValue)
+		if err == nil {
+			value = val
+		}
+	}
+	return value
 }
 
 //NewApisHandler creates new rest Handler instance