+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Implement upload image api [#10](https://github.com/rokwire/content-building-block/issues/10)
 
 ## [1.0.6] - 2021-06-23
 ### Fixed